feat: Add wordlist file reading functionality
The code changes include adding code to read a wordlist file in the `main.ipynb` notebook. This feature will allow the program to process a list of words from the file.

Note: The recent user commits and repository commits do not provide any additional information for generating the commit message.