feat: Add exercises for lists and loops
Added a new notebook file, fiche_1_3_exos_list.ipynb, which contains exercises for working with lists and loops. The exercises include creating a list of cubes of even integers, modifying the list by adding 1 to each value, replacing multiples of 3 with 0, creating a new list with repeated integers, adding the number of times an integer has appeared before to each entry, and more. These exercises provide practice in manipulating lists and performing calculations on their elements.